linux cmake project support
