+++ tinyphone/baseapp.cpp
@@ -12,6 +12,12 @@
 #include <boost/foreach.hpp>
 #include <math.h>
 
+#ifndef _WIN32
+#include <arpa/inet.h>
+#include <ifaddrs.h>
+#include <resolv.h>
+#endif
+
 using namespace std;
 using namespace pj;
 
@@ -29,6 +35,37 @@ namespace tp {
     std::string sipLogFile, httpLogFile;
     TinyPhoneHttpServer* tpHttpServer;
     tp::Endpoint ep;
+
+#ifndef _WIN32
+
+    std::vector<std::string> GetLocalDNSServers() {
+        std::vector <std::string> dnsServers;
+        // Get native iOS System Resolvers
+        res_ninit(&_res);
+        res_state res = &_res;
+
+        for (int i = 0; i < res->nscount; i++) {
+            sa_family_t family = res->nsaddr_list[i].sin_family;
+            int port = ntohs(res->nsaddr_list[i].sin_port);
+            if (family == AF_INET) { // IPV4 address
+                char str[INET_ADDRSTRLEN]; // String representation of address
+                inet_ntop(AF_INET, & (res->nsaddr_list[i].sin_addr.s_addr), str, INET_ADDRSTRLEN);
+                dnsServers.push_back(str);
+
+            } else if (family == AF_INET6) { // IPV6 address
+                char str[INET6_ADDRSTRLEN]; // String representation of address
+                inet_ntop(AF_INET6, &(res->nsaddr_list [i].sin_addr.s_addr), str, INET6_ADDRSTRLEN);
+            }
+        }
+#ifndef HAVE_RES_NDESTROY 
+        res_nclose(res);
+#else 
+        res_ndestroy(res);
+#endif
+        return dnsServers;
+    }
+
+#endif
 
     void InitPJSUAEndpoint(std::string logfile) {
         /* Create endpoint instance! */
@@ -155,5 +192,7 @@ namespace tp {
     TinyPhone* GetPhone(){
         return tp::tpHttpServer->tinyPhone;
     }
+
+
     
 }
